+++ packages/fhir_r6_auth/lib/src/authenticators/authenticators.dart
@@ -1,3 +1,0 @@
-export 'authenticator.dart';
-export 'mobile_authenticator.dart';
-export 'web_authenticator.dart';